dictum-core/pyproject.toml
[tool.poetry]
name = "dictum-core"
version = "0.1.13"
description = "Core library for Dictum"
authors = ["Mikhail Akimov"]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pydantic = "2.4.2"
lark = "^1.1.8"
PyYAML = "^6.0.1"
SQLAlchemy = "^2.0.23"
Jinja2 = "^3.0.1"
sqlparse = "^0.4.2"
Babel = "^2.9.1"
jupyter = "^1.0.0"
graphlib_backport = { version = "^1.0.3", python = "<3.9" }
toolz = "^0.12.0"
pandas = "^2.1.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
isort = "^5.9.3"
pytest = "^6.2.5"
pytest-cov = { extras = ["toml"], version = "^3.0.0" }
mkdocs-material = "^8.1.7"
mkdocs-jupyter = "^0.19.0"
black = "^22.3.0"
jupyter = "^1.0.0"

[tool.poetry.scripts]
dictum = "dictum_core.cli:app"

[tool.pytest.ini_options]
addopts = "--cov-report=xml --cov=dictum_core"

[tool.isort]
profile = "black"
known_first_party = ["dictum_core"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
